関数名の重複を修正
